Switch to qurgh pIqaD font
This font has more "classical" glyph shapes than the HaSta font.
The qurgh font was downloaded from http://hol.kag.org, which displayed
the following notice regarding the font:

All the fonts (KApIqaD, pIqaD, HaSta - TrueType TTF and Web woff) used on
this page are available <a href="pIqaDFontsNormalAndWeb.zip">here</a>.
These are free to use, but we ask that this page is mentioned if they are
used in a commercial product.