Display key legends above key when pressed
Scale and translate the button so that it goes directly upward, and apply
a complementary scaling and translation to the title label to keep it
from stretching out and to place it at the top of the stretched button.
The affine transform seems to be undone automatically when applied to
the title label, so achieve the translation using an edge inset instead.